+++ icetea/Tx.js
@@ -2,11 +2,10 @@ const { TxOp } = require('./enum')
 const { sha256 } = require('../icetea/helper/codec')
 
 module.exports = class {
-  // data is null or empty: normal tx
   // create contract
   // data = {
   //    op: 0,
-  //    src: "console.log('hello worl')"";
+  //    src: "console.log('hello world')";
   // }
   // call contract function
   // data = {
@@ -18,12 +17,17 @@ module.exports = class {
   // Some op in the future: set alias/options, vote, etc.
 
   constructor (from, to, value, fee, data, nonce) {
+
+    if (!from) {
+      throw new Error('Transaction "from" is required.')
+    }
+
     this.from = from || ''
     this.to = to || ''
     this.value = parseFloat(value) || 0
     this.fee = parseFloat(fee) || 0
     this.data = data || {}
-    this.nonce = nonce || Date.now()
+    this.nonce = nonce || Date.now() // FIXME
 
     if (this.value < 0 || this.fee < 0) {
       throw new Error('Value and fee cannot be negative.')
