+++ icetea/config/index.js
@@ -1,4 +1,5 @@
 module.exports = {
+  abciServerPort: 26658,
   feeCollector: '617BFqg1QhNtsJiNiWz9jGpsm5iAJKqWQBhhk36KjvUFqNkh47',
   initialBalances: [
     {
